sonic/api/parser.py
from result import Err, Ok, Result

from sonic.api.errors import MissingFieldsError
from sonic.service_layer import services


def parse_transaction(
    t: str,
) -> Result[services.AddTransactionRequest, MissingFieldsError]:
    pairs = t.split(";")
    hashmap = {}

    for pair in pairs:
        key, value = pair.split("=")
        hashmap[key] = value

    missing_fields = {
        "client_id",
        "transaction_timestamp",
        "value",
        "description",
    } - hashmap.keys()

    if len(missing_fields):
        return Err(MissingFieldsError(missing_fields))

    return Ok(
        services.AddTransactionRequest(
            client_id=hashmap["client_id"],
            timestamp=hashmap["transaction_timestamp"],
            value=hashmap["value"],
            description=hashmap["description"],
        )
    )
